Update getIdentifier and add getElement functions
Updated the getIdentifier function in the PjsDB class to accept a parameter for specifying the element type. Now, it retrieves the last used identifier specifically for the given element type in the XML file. This enhancement improves the accuracy of identifier retrieval for different types of elements.

Additionally, the getElement function has been added to the PjsDB class. It allows retrieving information from the XML file based on provided parameters. When only the $filename parameter is provided, the function now correctly returns all the information from the XML file. The implementation includes XPath queries to selectively retrieve elements and their attributes. The retrieved information is formatted and returned in JSON format, facilitating data consumption and integration.